LegalEase/requirements.txt
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.5.2
